# code/storage/app/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kanban_project_backend_api",
        "contact": {
            "name": "kent vejrup madsen",
            "url": "https://github.com/KentVejrupMadsen",
            "email": "kent.vejrup.madsen@protonmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/KentVejrupMadsen/Kanban-Project-Backend/blob/main/License.md"
        },
        "version": "1.0.0-alpha"
    },
    "servers": [
        {
            "url": "https://kanban.goalpioneers.com/api/1.0.0",
            "description": "public available server"
        },
        {
            "url": "http://localhost:8000/api/1.0.0",
            "description": "local development server"
        }
    ],
    "paths": {
        "/api/1.0.0/accounts/account/me": {
            "get": {
                "tags": [
                    "1.0.0",
                    "account",
                    "authentication"
                ],
                "operationId": "c0bab5220530bfa07223c74f6f214cb2",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "<<<JSON"
                            },
                            "application/xml": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/account/read": {
            "get": {
                "tags": [
                    "1.0.0",
                    "account"
                ],
                "operationId": "121077445940ac8019248752b6ede69c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/account/login": {
            "post": {
                "tags": [
                    "1.0.0",
                    "account",
                    "authentication"
                ],
                "operationId": "e832983ffa2b58622abe013f6731bc5e",
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/account/logout": {
            "get": {
                "tags": [
                    "1.0.0",
                    "account",
                    "authentication"
                ],
                "operationId": "cf3e4f718e9cb3d650efb0d62d647eec",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/account/create": {
            "post": {
                "tags": [
                    "1.0.0",
                    "account",
                    "authentication"
                ],
                "description": "",
                "externalDocs": {
                    "description": "",
                    "url": "https://github.com/KentVejrupMadsen/Kanban-Project-Backend"
                },
                "operationId": "aad533bcd63625a1775284f06134073a",
                "responses": {
                    "201": {
                        "description": "Account created",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "{\r\n\t\"username\": \"username\",\r\n\t\"bearer_token\": \"integer|bearer-token-string-herexxxxxx\"\r\n}"
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - an account already exist with the given parameters"
                    },
                    "540": {
                        "description": "content not found"
                    }
                },
                "deprecated": false
            }
        },
        "/api/1.0.0/accounts/account/update": {
            "patch": {
                "tags": [
                    "1.0.0",
                    "account"
                ],
                "operationId": "70dd13b0ed7285230395b950e6d0e0a3",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/account/delete": {
            "delete": {
                "tags": [
                    "1.0.0",
                    "account"
                ],
                "operationId": "f832cada551424587743343b2b5e6723",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
        "/api/1.0.0/accounts/account/verify": {
            "post": {
                "tags": [
                    "1.0.0",
                    "account",
                    "authentication"
                ],
                "operationId": "7c0f062fae9f153af1d3c7b6f80e6d9e",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/xml": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "content not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "account_logout": {
                "type": "http",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            },
            "account_update": {
                "type": "http",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            },
            "account_deletion": {
                "type": "http",
                "description": "",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}